refactor(inception score): rename file with the inception score metric (closes #74)
