+++ photosynthesis_metrics/__init__.py
@@ -3,8 +3,7 @@ from .msid import MSID
 from .fid import FID
 from .kid import KID
 from .tv import TVLoss, total_variation
-from .inception_score import IS, inception_score
 from .gmsd import GMSDLoss, MultiScaleGMSDLoss
 from .gs import GS
-from .inception_score import IS, inception_score
+from .isc import IS, inception_score
 from .vif import VIFLoss, vif_p
